Rename fasta_import.py to db_import.py
